chore: add normalize utils
