+++ list-sync/src/allowlist.ts
@@ -1,7 +1,7 @@
 import { readdirSync, readFileSync } from 'fs';
 import { join } from 'path';
 import { parse } from 'yaml';
-import { keccak256, toUtf8Bytes } from 'ethers';
+import { normalizeAddress, normalizeSelector } from './utils.js';
 
 export type SelectorState = 'OFF' | 'ON' | 'PERMANENTLY_ON';
 
@@ -102,38 +102,6 @@ function normalizeState(value: string, context: string): SelectorState {
     return value as SelectorState;
   }
   throw new Error(`Invalid allowlist state "${value}" (${context})`);
-}
-
-function normalizeAddress(value: string, context: string): string {
-  if (typeof value !== 'string') {
-    throw new Error(`Expected address string for ${context}`);
-  }
-  if (!value.startsWith('0x')) {
-    throw new Error(`Address missing 0x prefix for ${context}: ${value}`);
-  }
-  if (value.length !== 42) {
-    throw new Error(`Address must be 42 chars for ${context}: ${value}`);
-  }
-  return value.toLowerCase();
-}
-
-function normalizeSelector(value: string, context: string): { selector: string; signature?: string } {
-  if (typeof value !== 'string') {
-    throw new Error(`Expected selector string for ${context}`);
-  }
-
-  const trimmed = value.trim();
-
-  if (/^0x[0-9a-fA-F]{8}$/.test(trimmed)) {
-    return { selector: trimmed.toLowerCase() };
-  }
-
-  if (trimmed.includes('(') && trimmed.endsWith(')')) {
-    const hash = keccak256(toUtf8Bytes(trimmed));
-    return { selector: hash.slice(0, 10), signature: trimmed };
-  }
-
-  throw new Error(`Selector must be 4-byte hex or signature for ${context}: ${value}`);
 }
 
 export function loadAllowlistFile(path: string): AllowlistNormalized {
